Refactor help to an extension
